chore(refactor): introduce LockService
